requirements.txt
Pillow>=10.0.0
numpy>=1.21.0
pyyaml>=6.0
